==> Data_merging/1_Air_Quality_Data_merged.ipynb <==
import pandas as pd
import numpy as np
import gc

## Importing and merging preprocessed air pollutants data

# Import clean ozone data
ozone_df = pd.read_csv(r"C:\Users\stlva\Google Drive\Big blue data academy\main_branch\concatenated data\all_ozone_clean.csv")

ozone_df.info()

# Import clean SO2 data
SO2_df = pd.read_csv(r"C:\Users\stlva\Google Drive\Big blue data academy\main_branch\concatenated data\all_SO2_clean.csv")

SO2_df.info()

# Merge ozone and SO2
air_all_df = pd.merge(ozone_df, SO2_df, how='outer', on='measurement_id')
air_all_df.head()

# Deleting previously merged dataframes in order to save memory
del ozone_df
del SO2_df
gc.collect()

# Fill NA values in the left table, from the corresponing column in the right table.
for column in air_all_df.columns:
    if column.endswith('_x'):
        corresponding_column = column[:-2] + '_y'
        if corresponding_column in air_all_df.columns:
            air_all_df[column] = air_all_df[column].fillna(air_all_df[corresponding_column])

# Drop all columns from the right table, so that we don't have dupblicate columns
air_all_df = air_all_df.drop(columns=air_all_df.filter(regex='_y$').columns)

# Rename the columns
air_all_df.rename(columns={'State Code_x':'State Code', 'State Name_x':'State Name', 'County Code_x':'County Code', 'County Name_x':'County Name',
       'City Name_x':'City Name', 'Site Num_x':'Site Num', 'Datum_x':'Datum', 'Latitude_x':'Latitude', 'Longitude_x':'Longitude',
       'Date Local_x':'Date Local', 'CBSA Name_x':'CBSA Name', 'AQI_x':'AQI'},inplace=True)

air_all_df.head()

# Import clean NO2 data
NO2_df = pd.read_csv(r"C:\Users\stlva\Google Drive\Big blue data academy\main_branch\concatenated data\all_NO2_clean.csv")

NO2_df.info()

# Merge NO2 data with the previously merged table
air_all_df = pd.merge(air_all_df, NO2_df, how='outer', on='measurement_id')

# Deleting previously merged dataframes in order to save memory
del NO2_df
gc.collect()

# Fill NA values in the left table, from the corresponing column in the right table.
for column in air_all_df.columns:
    if column.endswith('_x'):
        corresponding_column = column[:-2] + '_y'
        if corresponding_column in air_all_df.columns:
            air_all_df[column] = air_all_df[column].fillna(air_all_df[corresponding_column])

# Drop all columns from the right table, so that we don't have dupblicate columns
air_all_df = air_all_df.drop(columns=air_all_df.filter(regex='_y$').columns)

# Rename the columns
air_all_df.rename(columns={'State Code_x':'State Code', 'State Name_x':'State Name', 'County Code_x':'County Code', 'County Name_x':'County Name',
       'City Name_x':'City Name', 'Site Num_x':'Site Num', 'Datum_x':'Datum', 'Latitude_x':'Latitude', 'Longitude_x':'Longitude',
       'Date Local_x':'Date Local', 'CBSA Name_x':'CBSA Name', 'AQI_x':'AQI'},inplace=True)

air_all_df.head()

# Import clean PM2_5 data
PM2_5_df = pd.read_csv(r"C:\Users\stlva\Google Drive\Big blue data academy\main_branch\concatenated data\all_PM2_5_clean.csv")

PM2_5_df.info()

# Merge PM2.5 data with the previously merged table
air_all_df = pd.merge(air_all_df, PM2_5_df, how='outer', on='measurement_id')

# Deleting previously merged dataframes in order to save memory
del PM2_5_df
gc.collect()

# Fill NA values in the left table, from the corresponing column in the right table.
for column in air_all_df.columns:
    if column.endswith('_x'):
        corresponding_column = column[:-2] + '_y'
        if corresponding_column in air_all_df.columns:
            air_all_df[column] = air_all_df[column].fillna(air_all_df[corresponding_column])

# Drop all columns from the right table, so that we don't have dupblicate columns
air_all_df = air_all_df.drop(columns=air_all_df.filter(regex='_y$').columns)

# Rename the columns
air_all_df.rename(columns={'State Code_x':'State Code', 'State Name_x':'State Name', 'County Code_x':'County Code', 'County Name_x':'County Name',
       'City Name_x':'City Name', 'Site Num_x':'Site Num', 'Datum_x':'Datum', 'Latitude_x':'Latitude', 'Longitude_x':'Longitude',
       'Date Local_x':'Date Local', 'CBSA Name_x':'CBSA Name', 'AQI_x':'AQI'},inplace=True)

air_all_df.head()

# Import clean PM10 data
PM10_df = pd.read_csv(r"C:\Users\stlva\Google Drive\Big blue data academy\main_branch\concatenated data\all_PM10_clean.csv")

PM10_df.info()

# Merge PM10 data with the previously merged table
air_all_df = pd.merge(air_all_df, PM10_df, how='outer', on='measurement_id')

# Deleting previously merged dataframes in order to save memory
del PM10_df
gc.collect()

# Fill NA values in the left table, from the corresponing column in the right table.
for column in air_all_df.columns:
    if column.endswith('_x'):
        corresponding_column = column[:-2] + '_y'
        if corresponding_column in air_all_df.columns:
            air_all_df[column] = air_all_df[column].fillna(air_all_df[corresponding_column])

# Drop all columns from the right table, so that we don't have dupblicate columns
air_all_df = air_all_df.drop(columns=air_all_df.filter(regex='_y$').columns)

# Rename the columns
air_all_df.rename(columns={'State Code_x':'State Code', 'State Name_x':'State Name', 'County Code_x':'County Code', 'County Name_x':'County Name',
       'City Name_x':'City Name', 'Site Num_x':'Site Num', 'Datum_x':'Datum', 'Latitude_x':'Latitude', 'Longitude_x':'Longitude',
       'Date Local_x':'Date Local', 'CBSA Name_x':'CBSA Name', 'AQI_x':'AQI'},inplace=True)

air_all_df.head()

# Import clean CO data
CO_df = pd.read_csv(r"C:\Users\stlva\Google Drive\Big blue data academy\main_branch\concatenated data\all_CO_clean.csv")

CO_df.info()

# Merge CO data with the previously merged table
air_all_df = pd.merge(air_all_df, CO_df, how='outer', on='measurement_id')

# Deleting previously merged dataframes in order to save memory
del CO_df
gc.collect()

# Fill NA values in the left table, from the corresponing column in the right table.
for column in air_all_df.columns:
    if column.endswith('_x'):
        corresponding_column = column[:-2] + '_y'
        if corresponding_column in air_all_df.columns:
            air_all_df[column] = air_all_df[column].fillna(air_all_df[corresponding_column])

# Drop all columns from the right table, so that we don't have dupblicate columns
air_all_df = air_all_df.drop(columns=air_all_df.filter(regex='_y$').columns)

# Rename the columns
air_all_df.rename(columns={'State Code_x':'State Code', 'State Name_x':'State Name', 'County Code_x':'County Code', 'County Name_x':'County Name',
       'City Name_x':'City Name', 'Site Num_x':'Site Num', 'Datum_x':'Datum', 'Latitude_x':'Latitude', 'Longitude_x':'Longitude',
       'Date Local_x':'Date Local', 'CBSA Name_x':'CBSA Name', 'AQI_x':'AQI'},inplace=True)

air_all_df.head()

air_all_df.columns

air_all_df['State Code'] = air_all_df['State Code'].astype(int).astype(str)
air_all_df['County Code'] = air_all_df['County Code'].astype(int).astype(str)
air_all_df['Site Num'] = air_all_df['Site Num'].astype(int).astype(str)


# Save the merged df to csv
air_all_df.to_csv("air_data_merged.csv",index=False)
