==> Data_merging/2_AQI_FILE.ipynb <==
import pandas as pd
import glob
import csv

## Importing csvs for all the parameters

ozone = pd.read_csv("all_ozone_clean.csv")

NO2 = pd.read_csv("all_NO2_clean.csv")

SO2 = pd.read_csv("all_SO2_clean.csv")

PM2_5 = pd.read_csv("all_PM2_5_clean.csv")

PM10 = pd.read_csv("all_PM10_clean.csv")

CO = pd.read_csv("all_CO_clean.csv")

## Keeping AQI from each parameter

selected_cols_ozone = ozone[['measurement_id', 'AQI']]
selected_cols_ozone

selected_cols_NO2 = NO2[['measurement_id', 'AQI']]
selected_cols_NO2

selected_cols_SO2 = SO2[['measurement_id', 'AQI']]
selected_cols_SO2

selected_cols_PM2_5 = PM2_5[['measurement_id', 'AQI']]
selected_cols_PM2_5

selected_cols_PM10 = PM10[['measurement_id', 'AQI']]
selected_cols_PM10

selected_cols_CO = CO[['measurement_id', 'AQI']]
selected_cols_CO

## Merging dataframes

merge1 = pd.merge(selected_cols_ozone, selected_cols_NO2, on='measurement_id',how='outer')
merge1

merge1.rename(columns={'AQI_x':'AQI_ozone', 'AQI_y':"AQI_NO2"}, inplace=True)
merge1

merge1 = pd.merge(merge1, selected_cols_SO2, on='measurement_id',how='outer')
merge1

merge1.rename(columns={'AQI':"AQI_SO2"}, inplace=True)
merge1

merge1 = pd.merge(merge1, selected_cols_CO, on='measurement_id', how='outer')
merge1

merge1.rename(columns={'AQI':"AQI_CO"}, inplace=True)
merge1

merge1 = pd.merge(merge1, selected_cols_PM2_5, on='measurement_id', how='outer')
merge1

merge1.rename(columns={'AQI':"AQI_PM2_5"}, inplace=True)
merge1

merge1 = pd.merge(merge1, selected_cols_PM10, on='measurement_id', how='outer')
merge1

merge1.rename(columns={'AQI':"AQI_PM10"}, inplace=True)
merge1

merge1.columns

included_cols = ['AQI_ozone', 'AQI_NO2', 'AQI_SO2', 'AQI_CO', 'AQI_PM2_5', 'AQI_PM10']

merge1['max_value'] = merge1[included_cols].max(axis=1)

merge1

## Creating a column for critical parameter of AQI

merge1['AQI_cr_param'] = merge1[included_cols].idxmax(axis=1)

merge1

## Keeping only AQI and AQI critical parameter

merge2 = merge1.drop(columns=['AQI_ozone','AQI_NO2','AQI_SO2','AQI_CO','AQI_PM2_5','AQI_PM10'])
merge2

merge2.to_csv("AQI_final_df.csv", index=False, encoding="utf-8")
