==> Data_merging/3_air_data_&_AQI.ipynb <==
import pandas as pd

import dask.dataframe as dd

## Read pollutants csv 

ddf = dd.read_csv("air_data_merged.csv", usecols=['State Code', 'State Name', 'County Code', 'County Name', 'City Name','Site Num', 'Datum', 'Latitude', 'Longitude', 'Date Local','CBSA Name','measurement_id'],low_memory=False)

# Perform operations on the Dask DataFrame
result = ddf.compute()
result.head()

## Read AQI csv

aqi = pd.read_csv("AQI_final_df.csv")

## Merging pollutants csv and AQI csv 

merge = pd.merge(result, aqi, on='measurement_id', how='outer')
merge.head()

merge.info()

# Checking for duplicates or errors

merge['measurement_id'].nunique()

result['measurement_id'].nunique()

aqi.info()

merge[merge.duplicated(keep=False)]

merge.to_csv("daily_AQI.csv", index=False, encoding='utf-8')
