==> Data_merging/4_Weather_Data_merged.ipynb <==
import pandas as pd
import numpy as np
import gc

## Importing and merging preprocessed weather data

# Import clean pressure data
pres_df = pd.read_csv("all_pressure_clean.csv")

pres_df.info()

# Import clean RH data
rh_df = pd.read_csv("all_rh_clean.csv", low_memory=False)

rh_df.info()

rh_df.isnull().sum()

# Merge pressure and RH
all_df = pd.merge(pres_df, rh_df, how='outer', on='measurement_id')

# Fill NA values in the left table, from the corresponing column in the right table.
for column in all_df.columns:
    if column.endswith('_x'):
        corresponding_column = column[:-2] + '_y'
        if corresponding_column in all_df.columns:
            all_df[column] = all_df[column].fillna(all_df[corresponding_column])

# Drop all columns from the right table, so that we don't have dupblicate columns
all_df = all_df.drop(columns=all_df.filter(regex='_y$').columns)

# Rename the columns
all_df.rename(columns={'State Code_x':'State Code', 'State Name_x':'State Name', 'County Code_x':'County Code', 'County Name_x':'County Name',
       'City Name_x':'City Name', 'Site Num_x':'Site Num', 'Datum_x':'Datum', 'Latitude_x':'Latitude', 'Longitude_x':'Longitude',
       'Date Local_x':'Date Local','CBSA Name_x':'CBSA Name'},inplace=True)

# Import clean temperature data
temp_df = pd.read_csv("all_temperature_clean.csv", low_memory=False)

# Merge temperature data with the previously merged table
all_df = pd.merge(all_df, temp_df, how='outer', on='measurement_id')

# Fill NA values in the left table, from the corresponing column in the right table.
for column in all_df.columns:
    if column.endswith('_x'):
        corresponding_column = column[:-2] + '_y'
        if corresponding_column in all_df.columns:
            all_df[column] = all_df[column].fillna(all_df[corresponding_column])

# Drop all columns from the right table, so that we don't have dupblicate columns
all_df = all_df.drop(columns=all_df.filter(regex='_y$').columns)

# Rename the columns
all_df.rename(columns={'State Code_x':'State Code', 'State Name_x':'State Name', 'County Code_x':'County Code', 'County Name_x':'County Name',
       'City Name_x':'City Name', 'Site Num_x':'Site Num', 'Datum_x':'Datum', 'Latitude_x':'Latitude', 'Longitude_x':'Longitude',
       'Date Local_x':'Date Local','CBSA Name_x':'CBSA Name'},inplace=True)

# Import clean wind data
wind_df = pd.read_csv("all_winds_clean.csv", low_memory=False)

# Merge wind data with the previously merged table
all_df = pd.merge(all_df, wind_df, how='outer', on='measurement_id')

# Fill NA values in the left table, from the corresponing column in the right table.
for column in all_df.columns:
    if column.endswith('_x'):
        corresponding_column = column[:-2] + '_y'
        if corresponding_column in all_df.columns:
            all_df[column] = all_df[column].fillna(all_df[corresponding_column])

# Drop all columns from the right table, so that we don't have dupblicate columns
all_df = all_df.drop(columns=all_df.filter(regex='_y$').columns)

# Rename the columns
all_df.rename(columns={'State Code_x':'State Code', 'State Name_x':'State Name', 'County Code_x':'County Code', 'County Name_x':'County Name',
       'City Name_x':'City Name', 'Site Num_x':'Site Num', 'Datum_x':'Datum', 'Latitude_x':'Latitude', 'Longitude_x':'Longitude',
       'Date Local_x':'Date Local','CBSA Name_x':'CBSA Name'},inplace=True)

all_df.columns

all_df['State Code'] = all_df['State Code'].astype(int).astype(str)
all_df['County Code'] = all_df['County Code'].astype(int).astype(str)
all_df['Site Num'] = all_df['Site Num'].astype(int).astype(str)

# Create a unique value for every station
all_df['station_id'] = all_df['State Code'].apply(str) + '_' + all_df['County Code'].apply(str) + '_' + all_df['Site Num'].apply(str)

# Save the merged df to csv
all_df.to_csv("weather_data.csv",index=False)
